Renamed DrawFlatColor to DrawFlatColored
