+++ amethyst_renderer/src/pass/flat_colored/interleaved.rs
@@ -1,5 +1,8 @@
 //! Simple flat forward drawing pass.
 
+use std::marker::PhantomData;
+
+use derivative::Derivative;
 use gfx::pso::buffer::ElemStride;
 use gfx_core::state::{Blend, ColorMask};
 use glsl_layout::Uniform;
@@ -10,35 +13,25 @@ use amethyst_core::{
     transform::GlobalTransform,
 };
 use amethyst_error::Error;
-use derivative::Derivative;
 
 use crate::{
     cam::{ActiveCamera, Camera},
     hidden::{Hidden, HiddenPropagate},
     mesh::{Mesh, MeshHandle},
     mtl::{Material, MaterialDefaults},
-    pass::{
-        skinning::{create_skinning_effect, setup_skinning_buffers},
-        util::{draw_mesh, get_camera, VertexArgs},
-    },
+    pass::util::{draw_mesh, get_camera, VertexArgs},
     pipe::{
         pass::{Pass, PassData},
         DepthMode, Effect, NewEffect,
     },
-    skinning::JointTransforms,
     tex::Texture,
     types::{Encoder, Factory},
-    vertex::{Attributes, Color, Position, Separate, VertexFormat},
+    vertex::{Color, Position, Query},
     visibility::Visibility,
     Rgba,
 };
 
 use super::*;
-
-static ATTRIBUTES: [Attributes<'static>; 2] = [
-    Separate::<Position>::ATTRIBUTES,
-    Separate::<Color>::ATTRIBUTES,
-];
 
 /// Draw mesh without lighting
 ///
@@ -49,25 +42,20 @@ static ATTRIBUTES: [Attributes<'static>; 2] = [
 ///
 /// * `V`: `VertexFormat`
 #[derive(Derivative, Clone, Debug, PartialEq)]
-#[derivative(Default(bound = "Self: Pass"))]
-pub struct DrawFlatColorSeparate {
-    skinning: bool,
+#[derivative(Default(bound = "V: Query<(Position, Color)>, Self: Pass"))]
+pub struct DrawFlatColored<V> {
+    _pd: PhantomData<V>,
     transparency: Option<(ColorMask, Blend, Option<DepthMode>)>,
 }
 
-impl DrawFlatColorSeparate
+impl<V> DrawFlatColored<V>
 where
+    V: Query<(Position, Color)>,
     Self: Pass,
 {
-    /// Create instance of `DrawFlatColor` pass
+    /// Create instance of `DrawFlatColored` pass
     pub fn new() -> Self {
         Default::default()
-    }
-
-    /// Enable vertex skinning
-    pub fn with_vertex_skinning(mut self) -> Self {
-        self.skinning = true;
-        self
     }
 
     /// Enable transparency
@@ -82,7 +70,10 @@ where
     }
 }
 
-impl<'a> PassData<'a> for DrawFlatColorSeparate {
+impl<'a, V> PassData<'a> for DrawFlatColored<V>
+where
+    V: Query<(Position, Color)>,
+{
     type Data = (
         Read<'a, ActiveCamera>,
         ReadStorage<'a, Camera>,
@@ -95,38 +86,24 @@ impl<'a> PassData<'a> for DrawFlatColorSeparate {
         ReadStorage<'a, MeshHandle>,
         ReadStorage<'a, Material>,
         ReadStorage<'a, GlobalTransform>,
-        ReadStorage<'a, JointTransforms>,
         ReadStorage<'a, Rgba>,
     );
 }
 
-impl Pass for DrawFlatColorSeparate {
+impl<V> Pass for DrawFlatColored<V>
+where
+    V: Query<(Position, Color)>,
+{
     fn compile(&mut self, effect: NewEffect<'_>) -> Result<Effect, Error> {
         use std::mem;
-        let mut builder = if self.skinning {
-            create_skinning_effect(effect, FRAG_SRC)
-        } else {
-            effect.simple(VERT_SRC, FRAG_SRC)
-        };
+        let mut builder = effect.simple(VERT_SRC, FRAG_SRC);
         builder
-            .with_raw_vertex_buffer(
-                Separate::<Position>::ATTRIBUTES,
-                Separate::<Position>::size() as ElemStride,
-                0,
+            .with_raw_constant_buffer(
+                "VertexArgs",
+                mem::size_of::<<VertexArgs as Uniform>::Std140>(),
+                1,
             )
-            .with_raw_vertex_buffer(
-                Separate::<Color>::ATTRIBUTES,
-                Separate::<Color>::size() as ElemStride,
-                0,
-            );
-        if self.skinning {
-            setup_skinning_buffers(&mut builder);
-        }
-        builder.with_raw_constant_buffer(
-            "VertexArgs",
-            mem::size_of::<<VertexArgs as Uniform>::Std140>(),
-            1,
-        );
+            .with_raw_vertex_buffer(V::QUERIED_ATTRIBUTES, V::size() as ElemStride, 0);
         match self.transparency {
             Some((mask, blend, depth)) => builder.with_blended_output("color", mask, blend, depth),
             None => builder.with_output("color", Some(DepthMode::LessEqualWrite)),
@@ -151,7 +128,6 @@ impl Pass for DrawFlatColorSeparate {
             mesh,
             material,
             global,
-            joints,
             rgba,
         ): <Self as PassData<'a>>::Data,
     ) {
@@ -159,8 +135,7 @@ impl Pass for DrawFlatColorSeparate {
 
         match visibility {
             None => {
-                for (joint, mesh, material, global, rgba, _, _) in (
-                    joints.maybe(),
+                for (mesh, material, global, rgba, _, _) in (
                     &mesh,
                     &material,
                     &global,
@@ -173,23 +148,22 @@ impl Pass for DrawFlatColorSeparate {
                     draw_mesh(
                         encoder,
                         effect,
-                        self.skinning,
+                        false,
                         mesh_storage.get(mesh),
-                        joint,
+                        None,
                         &tex_storage,
                         Some(material),
                         &material_defaults,
                         rgba,
                         camera,
                         Some(global),
-                        &ATTRIBUTES,
+                        &[V::QUERIED_ATTRIBUTES],
                         &[],
                     );
                 }
             }
             Some(ref visibility) => {
-                for (joint, mesh, material, global, rgba, _) in (
-                    joints.maybe(),
+                for (mesh, material, global, rgba, _) in (
                     &mesh,
                     &material,
                     &global,
@@ -201,16 +175,16 @@ impl Pass for DrawFlatColorSeparate {
                     draw_mesh(
                         encoder,
                         effect,
-                        self.skinning,
+                        false,
                         mesh_storage.get(mesh),
-                        joint,
+                        None,
                         &tex_storage,
                         Some(material),
                         &material_defaults,
                         rgba,
                         camera,
                         Some(global),
-                        &ATTRIBUTES,
+                        &[V::QUERIED_ATTRIBUTES],
                         &[],
                     );
                 }
@@ -220,16 +194,16 @@ impl Pass for DrawFlatColorSeparate {
                         draw_mesh(
                             encoder,
                             effect,
-                            self.skinning,
+                            false,
                             mesh_storage.get(mesh),
-                            joints.get(*entity),
+                            None,
                             &tex_storage,
                             material.get(*entity),
                             &material_defaults,
                             rgba.get(*entity),
                             camera,
                             global.get(*entity),
-                            &ATTRIBUTES,
+                            &[V::QUERIED_ATTRIBUTES],
                             &[],
                         );
                     }
