+++ amethyst_renderer/src/pass/mod.rs
@@ -4,7 +4,7 @@ pub use self::{
     debug_lines::*,
     flat::*,
     flat2d::*,
-    flat_color::*,
+    flat_colored::*,
     pbm::*,
     shaded::*,
     skinning::set_skinning_buffers,
@@ -15,7 +15,7 @@ pub use self::{
 mod debug_lines;
 mod flat;
 mod flat2d;
-mod flat_color;
+mod flat_colored;
 mod pbm;
 mod shaded;
 mod shaded_util;
