+++ amethyst_renderer/src/pass/shaders/fragment/flat_colored.glsl
