+++ amethyst_renderer/src/pass/shaders/vertex/flat_colored.glsl
